fix(admin-ui): resolve react-refresh and remaining no-explicit-any
- Extract STEP_TYPE_META into its own module so FlowStepNode only exports
  a component (react-refresh/only-export-components).
- Disable react-refresh on co-located context hooks and the test-only
  re-export helper, with rationale comments.
- ClientScopeCreatePage: type mutation onError as Error.
- RealmListPage: use getErrorMessage() for the caught error and avoid the
  any cast on the import result.
